started rework of how snippets are handled
